+++ crates/forge-runner/src/lib.rs
@@ -7,21 +7,16 @@ use crate::test_crate_summary::TestCrateSummary;
 use anyhow::{anyhow, Result};
 
 use cairo_lang_runner::RunnerError;
-use cairo_lang_sierra::extensions::core::{CoreLibfunc, CoreType};
 use cairo_lang_sierra::ids::ConcreteTypeId;
 use cairo_lang_sierra::program::{Function, Program};
-use cairo_lang_sierra::program_registry::ProgramRegistry;
 use cairo_lang_sierra_to_casm::compiler::CairoProgram;
 use cairo_lang_sierra_to_casm::metadata::{calc_metadata, MetadataComputationConfig};
-use cairo_lang_sierra_type_size::get_type_size_map;
 use camino::Utf8PathBuf;
 
 use contracts_data::ContractsData;
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
 
-use once_cell::sync::Lazy;
-use running::TestDetails;
 use smol_str::SmolStr;
 use trace_data::save_trace_data;
 
@@ -46,18 +41,16 @@ mod running;
 
 pub const CACHE_DIR: &str = ".snfoundry_cache";
 
-pub static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
-    vec![
-        "Pedersen",
-        "RangeCheck",
-        "Bitwise",
-        "EcOp",
-        "Poseidon",
-        "SegmentArena",
-        "GasBuiltin",
-        "System",
-    ]
-});
+pub const BUILTINS: [&str; 8] = [
+    "Pedersen",
+    "RangeCheck",
+    "Bitwise",
+    "EcOp",
+    "Poseidon",
+    "SegmentArena",
+    "GasBuiltin",
+    "System",
+];
 
 /// Configuration of the test runner
 #[derive(Debug, PartialEq)]
@@ -69,6 +62,7 @@ pub struct RunnerConfig {
     pub fuzzer_seed: u64,
     pub detailed_resources: bool,
     pub save_trace_data: bool,
+    pub max_n_steps: Option<u32>,
 }
 
 impl RunnerConfig {
@@ -81,6 +75,7 @@ impl RunnerConfig {
         fuzzer_seed: u64,
         detailed_resources: bool,
         save_trace_data: bool,
+        max_n_steps: Option<u32>,
     ) -> Self {
         Self {
             workspace_root,
@@ -89,6 +84,7 @@ impl RunnerConfig {
             fuzzer_seed,
             detailed_resources,
             save_trace_data,
+            max_n_steps,
         }
     }
 }
@@ -133,55 +129,6 @@ pub trait TestCaseFilter {
 pub enum TestCrateRunResult {
     Ok(TestCrateSummary),
     Interrupted(TestCrateSummary),
-}
-
-fn build_test_details(test_name: &str, sierra_program: &Program) -> TestDetails {
-    let sierra_program_registry =
-        ProgramRegistry::<CoreType, CoreLibfunc>::new(sierra_program).unwrap();
-    let type_sizes = get_type_size_map(sierra_program, &sierra_program_registry).unwrap();
-    let func = sierra_program
-        .funcs
-        .iter()
-        .find(|f| f.id.debug_name.clone().unwrap().ends_with(test_name))
-        .unwrap();
-
-    let parameter_types = func
-        .signature
-        .param_types
-        .iter()
-        .map(|pt| {
-            let td = sierra_program
-                .type_declarations
-                .iter()
-                .find(|td| &td.id == pt)
-                .unwrap();
-            let generic_id = &td.long_id.generic_id;
-            let size = type_sizes[&td.id];
-            (generic_id.clone(), size)
-        })
-        .collect::<Vec<_>>();
-
-    let return_types = func
-        .signature
-        .ret_types
-        .iter()
-        .map(|pt| {
-            let td = sierra_program
-                .type_declarations
-                .iter()
-                .find(|td| &td.id == pt)
-                .unwrap();
-            let generic_id = &td.long_id.generic_id;
-            let size = type_sizes[&td.id];
-            (generic_id.clone(), size)
-        })
-        .collect::<Vec<_>>();
-
-    TestDetails {
-        entry_point_offset: func.entry_point.0,
-        parameter_types,
-        return_types,
-    }
 }
 
 fn compile_sierra_to_casm(sierra_program: &Program) -> CairoProgram {
@@ -233,13 +180,11 @@ pub async fn run_tests_from_crate(
 
         let case = Arc::new(case.clone());
         let args: Vec<ConcreteTypeId> = args.into_iter().cloned().collect();
-        let test_details = Arc::new(build_test_details(&case.name, sierra_program));
 
         tasks.push(choose_test_strategy_and_run(
             args,
             case.clone(),
             casm_program.clone(),
-            test_details.clone(),
             runner_config.clone(),
             runner_params.clone(),
             send.clone(),
@@ -285,36 +230,20 @@ fn choose_test_strategy_and_run(
     args: Vec<ConcreteTypeId>,
     case: Arc<TestCaseRunnable>,
     casm_program: Arc<CairoProgram>,
-    test_details: Arc<TestDetails>,
     runner_config: Arc<RunnerConfig>,
     runner_params: Arc<RunnerParams>,
     send: Sender<()>,
 ) -> JoinHandle<Result<AnyTestCaseSummary>> {
     if args.is_empty() {
         tokio::task::spawn(async move {
-            let res = run_test(
-                case,
-                casm_program,
-                test_details,
-                runner_config,
-                runner_params,
-                send,
-            )
-            .await??;
+            let res = run_test(case, casm_program, runner_config, runner_params, send).await??;
             Ok(AnyTestCaseSummary::Single(res))
         })
     } else {
         tokio::task::spawn(async move {
-            let res = run_with_fuzzing(
-                args,
-                case,
-                casm_program,
-                test_details,
-                runner_config,
-                runner_params,
-                send,
-            )
-            .await??;
+            let res =
+                run_with_fuzzing(args, case, casm_program, runner_config, runner_params, send)
+                    .await??;
             Ok(AnyTestCaseSummary::Fuzzing(res))
         })
     }
@@ -324,7 +253,6 @@ fn run_with_fuzzing(
     args: Vec<ConcreteTypeId>,
     case: Arc<TestCaseRunnable>,
     casm_program: Arc<CairoProgram>,
-    test_details: Arc<TestDetails>,
     runner_config: Arc<RunnerConfig>,
     runner_params: Arc<RunnerParams>,
     send: Sender<()>,
@@ -363,7 +291,6 @@ fn run_with_fuzzing(
                 args,
                 case.clone(),
                 casm_program.clone(),
-                test_details.clone(),
                 runner_config.clone(),
                 runner_params.clone(),
                 send.clone(),