+++ crates/forge/tests/e2e/running.rs
@@ -79,7 +79,7 @@ fn simple_package_with_git_dependency() {
             casm = true
 
             [dependencies]
-            starknet = "2.5.0"
+            starknet = "2.5.4"
             snforge_std = {{ git = "https://github.com/{}", branch = "{}" }}
             "#,
             remote_url,
@@ -755,7 +755,7 @@ fn init_new_project_test() {
 
             [dependencies]
             snforge_std = {{ git = "https://github.com/foundry-rs/starknet-foundry", tag = "v{}" }}
-            starknet = "2.5.0"
+            starknet = "2.5.4"
 
             [[target.starknet-contract]]
             casm = true
@@ -778,7 +778,7 @@ fn init_new_project_test() {
         casm = true
 
         [dependencies]
-        starknet = "2.5.0"
+        starknet = "2.5.4"
         snforge_std = {{ git = "https://github.com/{}", branch = "{}" }}
         "#,
             remote_url,
@@ -937,7 +937,7 @@ fn incompatible_snforge_std_version_warning() {
         output,
         indoc! {r"
         [..]Updating git repository https://github.com/foundry-rs/starknet-foundry
-        [WARNING] Package snforge_std version does not meet the recommended version requirement =0.17.1, [..]
+        [WARNING] Package snforge_std version does not meet the recommended version requirement =0.18.0, [..]
         [..]Compiling[..]
         [..]Finished[..]