+++ crates/sncast/src/state.rs
@@ -137,7 +137,7 @@ fn verify_version(version: u8) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::response::structs::Hex;
+    use crate::response::structs::Felt;
     use crate::state::ScriptTransactionOutput::ErrorResponse;
     use camino::Utf8PathBuf;
     use tempfile::TempDir;
@@ -229,8 +229,8 @@ mod tests {
         let transaction = ScriptTransactionEntry {
             name: "declare".to_string(),
             output: ScriptTransactionOutput::DeclareResponse(DeclareResponse {
-                class_hash: Hex("0x123".parse().unwrap()),
-                transaction_hash: Hex("0x321".parse().unwrap()),
+                class_hash: Felt("0x123".parse().unwrap()),
+                transaction_hash: Felt("0x321".parse().unwrap()),
             }),
             status: ScriptTransactionStatus::Success,
             timestamp: 0,
@@ -265,8 +265,8 @@ mod tests {
         let transaction1 = ScriptTransactionEntry {
             name: "declare".to_string(),
             output: ScriptTransactionOutput::DeclareResponse(DeclareResponse {
-                class_hash: Hex("0x1".parse().unwrap()),
-                transaction_hash: Hex("0x2".parse().unwrap()),
+                class_hash: Felt("0x1".parse().unwrap()),
+                transaction_hash: Felt("0x2".parse().unwrap()),
             }),
             status: ScriptTransactionStatus::Success,
             timestamp: 0,
@@ -279,7 +279,7 @@ mod tests {
         let transaction2 = ScriptTransactionEntry {
             name: "invoke".to_string(),
             output: ScriptTransactionOutput::InvokeResponse(InvokeResponse {
-                transaction_hash: Hex("0x3".parse().unwrap()),
+                transaction_hash: Felt("0x3".parse().unwrap()),
             }),
             status: ScriptTransactionStatus::Success,
             timestamp: 1,
@@ -334,8 +334,8 @@ mod tests {
         let transaction1 = ScriptTransactionEntry {
             name: "declare".to_string(),
             output: ScriptTransactionOutput::DeclareResponse(DeclareResponse {
-                class_hash: Hex("0x1".parse().unwrap()),
-                transaction_hash: Hex("0x2".parse().unwrap()),
+                class_hash: Felt("0x1".parse().unwrap()),
+                transaction_hash: Felt("0x2".parse().unwrap()),
             }),
             status: ScriptTransactionStatus::Success,
             timestamp: 2,
@@ -347,7 +347,7 @@ mod tests {
         let transaction2 = ScriptTransactionEntry {
             name: "invoke".to_string(),
             output: ScriptTransactionOutput::InvokeResponse(InvokeResponse {
-                transaction_hash: Hex("0x3".parse().unwrap()),
+                transaction_hash: Felt("0x3".parse().unwrap()),
             }),
             status: ScriptTransactionStatus::Success,
             timestamp: 3,