+++ crates/sncast/tests/e2e/declare.rs
@@ -381,7 +381,7 @@ async fn test_no_scarb_profile() {
     let snapbox = runner(&args).current_dir(contract_path.path());
     snapbox.assert().success().stdout_matches(indoc! {r"
         ...
-        Warning: Profile profile5 does not exist in scarb, using default 'dev' profile.
+        [WARNING] Profile profile5 does not exist in scarb, using default 'dev' profile.
         command: declare
         class_hash: [..]
         transaction_hash: [..]