+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.0] - 2024-02-21
+
 ### Forge
 
 #### Added
 
 - contract names to call trace
+- `--max-n-steps` argument that allows setting own steps limit
+
+#### Changed
+
+- Unknown entry point error when calling a contract counts as a panic
 
 #### Fixed
 
@@ -22,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `script init` command to generate a template file structure for deployment scripts
+- Warning is emitted when executing sncast commands if the node's JSON-RPC version is incompatible
 
 #### Changed